requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
pydantic==2.7.0
python-multipart==0.0.9
openai==1.54.4
httpx==0.27.0
beautifulsoup4==4.12.3
selectolax==0.3.21
PyPDF2==3.0.1
pdfminer.six==20231228
pytesseract==0.3.10
pdf2image==1.17.0
zipstream-ng==1.7.1
aiofiles==24.1.0